refactor: package renaming core to common
