+++ src/main/java/com/onmoim/server/common/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.onmoim.server.core.exception;
+package com.onmoim.server.common.exception;
 
 import lombok.Getter;
