+++ src/main/java/com/onmoim/server/common/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.onmoim.server.core.exception;
+package com.onmoim.server.common.exception;
 
 import static org.springframework.http.HttpStatus.*;
