+++ src/main/java/com/onmoim/server/common/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.onmoim.server.core.exception;
+package com.onmoim.server.common.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-import com.onmoim.server.core.response.ResponseHandler;
+import com.onmoim.server.common.response.ResponseHandler;
 
 import lombok.extern.slf4j.Slf4j;
